refactor: extract link create command
